validation and error handeling

+++ backend/app/Http/Controllers/ClassesController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Classes;
 use Illuminate\Http\Request;
-
+use Illuminate\Support\Facades\Validator;
 class ClassesController extends Controller
 {
     public function __construct()
@@ -13,6 +13,14 @@ class ClassesController extends Controller
     }
     public function AddClass(Request $request)
     {
+        $validated=Validator::make($request->all(), [
+            'name' => 'required|string',
+            'floor' => 'required|string',
+            'color' => 'required|string|min:6|min:6'        
+        ]);
+        if($validated->fails()){
+            return response()->json(["message"=>$validated->errors()], 422);
+        }
         $class = new Classes();
         $name = $request->input('name');
         $floor = $request->input('floor');
@@ -21,7 +29,7 @@ class ClassesController extends Controller
         $class->save();
         return response()->json([
             "message" => "Class added successfully"
-        ]);
+        ], 201);
     }
 
     public function  GetClass(Request $request)
@@ -29,29 +37,53 @@ class ClassesController extends Controller
         $class = Classes::paginate(5);
         return response()->json([
             "message" => $class
-        ]);
+        ], 200);
     }
 
     public function getClassById(Request $request, $id)
     {
         $Classes = Classes::where('id', $id)->get();
+        if($Classes->isEmpty()){
+            return response()->json([
+                'message' => "Class doesn't exists",
+         
+            ], 400);
+        }
         return response()->json([
             "message" => $Classes
-        ]);
+        ], 200);
     }
 
     public function deleteClass(Request $request, $id)
     {
         $Classes = Classes::find($id);
+        if(empty($Classes)){
+            return response()->json([
+                'message' => "Class doesn't exists",
+            ], 400);
+        }
         $Classes->delete();
         return response()->json([
             "message" => "Classes Deleted Successfully!"
-        ]);
+        ], 200);
     }
 
     public function updateClass(Request $request, $id)
     {
+        $validated=Validator::make($request->all(), [
+            'name' => 'string',
+            'floor' => 'string',
+            'color' => 'string|min:6|min:6'
+        ]);
+        if($validated->fails()){
+            return response()->json(["message"=>$validated->errors()], 422);
+        }
         $Classes = Classes::find($id);
+        if(empty($Classes)){
+            return response()->json([
+                'message' => "Class doesn't exists",
+            ], 400);
+        }
         $inputs = $request->except("_method");
         if ($inputs) {
             $Classes->update($inputs);
@@ -59,6 +91,6 @@ class ClassesController extends Controller
         return response()->json([
             "message" => "Classes updated successfully",
             "Classes" => $Classes
-        ]);
+        ], 200);
     }
 }
